requirements.txt
ffmpeg
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
django-filter
pillow
dj_database_url
gunicorn
firebase-admin